GameOfLife completed.

+++ src/cs351/presets/RandomGrid.java
@@ -1,13 +1,24 @@
 package cs351.presets;
 
 import cs351.lab4.SimulationEngine;
-
 import java.util.Random;
 
+/**
+ * Sets the initial grid state to be randomized such that each cell
+ * has approximately a 50% chance of being alive.
+ *
+ * @author Justin Hall
+ */
 public class RandomGrid implements Preset
 {
   private final Random RAND = new Random();
 
+  /**
+   * At each cell, calls nextInt(100) and if that value is > 50, the cell
+   * is set to alive.
+   *
+   * @param engine SimulationEngine object to use
+   */
   @Override
   public void setInitialEngineState(SimulationEngine engine)
   {
@@ -15,8 +26,8 @@ public class RandomGrid implements Preset
     {
       for (int y = 0; y < engine.getWorldHeight(); y++)
       {
-        if (RAND.nextInt(100) > 50) engine.setAge(x, y, (byte)1);
-        else engine.setAge(x, y, (byte)0);
+        if (RAND.nextInt(100) > 50) engine.setAge(x, y, 1);
+        else engine.setAge(x, y, 0);
       }
     }
   }
